fix: geno/pheno missing

+++ app/Services/PromptService.php
@@ -207,15 +207,14 @@ class PromptService extends Service
             }
 
             $prompt = Prompt::create(Arr::only($data, ['prompt_category_id', 'name', 'summary', 'description', 'parsed_description', 'is_active', 'start_at', 'end_at', 'hide_before_start', 'hide_after_end', 'has_image', 'prefix', 'hide_submissions', 'level_req', 'limit', 'limit_period', 'limit_character', 'staff_only']));
-            
+
             $prompt->expreward()->create([
-                'prompt_id' => $prompt->id,
-                'user_exp'  => $data['user_exp'],
-                'user_points'  => $data['user_points'],
-                'chara_exp'  => $data['chara_exp'],
+                'prompt_id'     => $prompt->id,
+                'user_exp'      => $data['user_exp'],
+                'user_points'   => $data['user_points'],
+                'chara_exp'     => $data['chara_exp'],
                 'chara_points'  => $data['chara_points'],
             ]);
-
 
             if ($image) {
                 $this->handleImage($image, $prompt->imagePath, $prompt->imageFileName);
@@ -277,20 +276,19 @@ class PromptService extends Service
 
             $prompt->update(Arr::only($data, ['prompt_category_id', 'name', 'summary', 'description', 'parsed_description', 'is_active', 'start_at', 'end_at', 'hide_before_start', 'hide_after_end', 'has_image', 'prefix', 'hide_submissions', 'level_req', 'limit', 'limit_period', 'limit_character', 'staff_only']));
 
-            if($prompt->expreward) {
+            if ($prompt->expreward) {
                 $prompt->expreward()->update([
-                    'user_exp'  => $data['user_exp'],
-                    'user_points'  => $data['user_points'],
-                    'chara_exp'  => $data['chara_exp'],
+                    'user_exp'      => $data['user_exp'],
+                    'user_points'   => $data['user_points'],
+                    'chara_exp'     => $data['chara_exp'],
                     'chara_points'  => $data['chara_points'],
                 ]);
-            }
-            else {
+            } else {
                 $prompt->expreward()->create([
-                    'prompt_id' => $prompt->id,
-                    'user_exp'  => $data['user_exp'],
-                    'user_points'  => $data['user_points'],
-                    'chara_exp'  => $data['chara_exp'],
+                    'prompt_id'     => $prompt->id,
+                    'user_exp'      => $data['user_exp'],
+                    'user_points'   => $data['user_points'],
+                    'chara_exp'     => $data['chara_exp'],
                     'chara_points'  => $data['chara_points'],
                 ]);
             }
@@ -329,6 +327,39 @@ class PromptService extends Service
             }
 
             $prompt->rewards()->delete();
+            if ($prompt->has_image) {
+                $this->deleteImage($prompt->imagePath, $prompt->imageFileName);
+            }
+            $prompt->delete();
+
+            return $this->commitReturn(true);
+        } catch (\Exception $e) {
+            $this->setError('error', $e->getMessage());
+        }
+
+        return $this->rollbackReturn(false);
+    }
+
+    /**
+     * Deletes a prompt.
+     *
+     * @param \App\Models\Prompt\Prompt $prompt
+     *
+     * @return bool
+     */
+    public function deletePrompt($prompt)
+    {
+        DB::beginTransaction();
+
+        try {
+            // Check first if the category is currently in use
+            if (Submission::where('prompt_id', $prompt->id)->exists()) {
+                throw new \Exception('A submission under this prompt exists. Deleting the prompt will break the submission page - consider setting the prompt to be not active instead.');
+            }
+
+            $prompt->rewards()->delete();
+            $prompt->expreward()->delete();
+            $prompt->skills()->delete();
             if ($prompt->has_image) {
                 $this->deleteImage($prompt->imagePath, $prompt->imageFileName);
             }
@@ -383,12 +414,24 @@ class PromptService extends Service
             $data['parsed_description'] = parse($data['description']);
         }
 
-        if(!isset($data['hide_before_start'])) $data['hide_before_start'] = 0;
-        if(!isset($data['hide_after_end'])) $data['hide_after_end'] = 0;
-        if(!isset($data['is_active'])) $data['is_active'] = 0;
-        if(!isset($data['level_check'])) $data['level_req'] = null;
-        if(!isset($data['limit_character'])) $data['limit_character'] = null;
-        if(!isset($data['staff_only'])) $data['staff_only'] = null;
+        if (!isset($data['hide_before_start'])) {
+            $data['hide_before_start'] = 0;
+        }
+        if (!isset($data['hide_after_end'])) {
+            $data['hide_after_end'] = 0;
+        }
+        if (!isset($data['is_active'])) {
+            $data['is_active'] = 0;
+        }
+        if (!isset($data['level_check'])) {
+            $data['level_req'] = null;
+        }
+        if (!isset($data['limit_character'])) {
+            $data['limit_character'] = null;
+        }
+        if (!isset($data['staff_only'])) {
+            $data['staff_only'] = null;
+        }
         if (isset($data['remove_image'])) {
             if ($prompt && $prompt->has_image && $data['remove_image']) {
                 $data['has_image'] = 0;
@@ -426,51 +469,22 @@ class PromptService extends Service
     /**
      * Processes user input for creating/updating prompt skill rewards.
      *
-     * @param  array                      $data
-     * @param  \App\Models\Prompt\Prompt  $prompt
+     * @param array                     $data
+     * @param \App\Models\Prompt\Prompt $prompt
      */
     private function populateSkills($data, $prompt)
     {
         // Clear the old skills...
         $prompt->skills()->delete();
 
-        if(isset($data['skill_id'])) {
-            foreach($data['skill_id'] as $key => $type)
-            {
+        if (isset($data['skill_id'])) {
+            foreach ($data['skill_id'] as $key => $type) {
                 PromptSkill::create([
                     'prompt_id'       => $prompt->id,
-                    'skill_id' => $type,
+                    'skill_id'        => $type,
                     'quantity'        => $data['skill_quantity'][$key],
                 ]);
             }
         }
-    }
-
-
-    /**
-     * Deletes a prompt.
-     *
-     * @param  \App\Models\Prompt\Prompt  $prompt
-     * @return bool
-     */
-    public function deletePrompt($prompt)
-    {
-        DB::beginTransaction();
-
-        try {
-            // Check first if the category is currently in use
-            if(Submission::where('prompt_id', $prompt->id)->exists()) throw new \Exception("A submission under this prompt exists. Deleting the prompt will break the submission page - consider setting the prompt to be not active instead.");
-
-            $prompt->rewards()->delete();
-            $prompt->expreward()->delete();
-            $prompt->skills()->delete();
-            if($prompt->has_image) $this->deleteImage($prompt->imagePath, $prompt->imageFileName);
-            $prompt->delete();
-
-            return $this->commitReturn(true);
-        } catch(\Exception $e) {
-            $this->setError('error', $e->getMessage());
-        }
-        return $this->rollbackReturn(false);
     }
 }
